requirements-test.txt
# Test dependencies for scitrera-app-framework
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: for testing env file functionality
python-dotenv>=1.0.0
